requirements.txt
# Harvest MVP Requirements
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# HTTP Client (for Ollama)
requests>=2.31.0

# Environment Configuration
python-dotenv>=1.0.0

# Optional: OpenAI fallback
# openai>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.0.0
isort>=5.12.0
